a little update on collision box

+++ GAME1017_Template_W01/Tile.h
@@ -4,7 +4,7 @@ class Tile : public Sprite {
 public:
 	Tile(SDL_Rect s, SDL_FRect d, const char* p, std::string k) : Sprite(s, d, p, k)
 	{
-		m_collisionBox = { (int)m_dst.x, (int)m_dst.y,(int)m_dst.w,(int)m_dst.h };
+		m_collisionBox = { m_dst.x, m_dst.y, m_dst.w, m_dst.h };
 	}
 	void Update() override;
 private:
